% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Bosnia.R
\docType{data}
\name{Bosnia}
\alias{Bosnia}
\title{My dataset from CSV}
\format{
A data frame with the following variables:
\describe{
\item{Income_0B}{Household income for the control group before the experiment}
\item{Income_1B}{Household income for the treatment group before the experiment}
\item{Income_0F}{Household income for the control group after the experiment}
\item{Income_1F}{Household income for the treatment group after the experiment}
}
}
\usage{
data(Bosnia)
}
\description{
This dataset was imported from a CSV file and included in the IPAG package for demonstration. Data are taken from the article by
Augsburg, B., De Haas, R., Harmgart, H., & Meghir, C. (2015).
The impacts of microcredit: Evidence from Bosnia and Herzegovina.
American Economic Journal: Applied Economics, 7(1), 183-203.
}
\details{
\doi{DOI: 10.1257/app.20130272}
}
\keyword{datasets}
