% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{Confidence interval for a mean}
\usage{
mean_ci(x, level = 0.99, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{level}{Confidence level (default 0.99)}

\item{na.rm}{Remove NA values}
}
\value{
Object of class 'mean_ci'
}
\description{
Confidence interval for a mean
}
\examples{
x <- c(4.2, 5.1, 6.3, 5.8, 4.9)
mean_ci(x)
mean_ci(x, level = 0.95)
}
