% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsratio_ci.R
\name{oddsratio_ci}
\alias{oddsratio_ci}
\title{Confidence interval for odds ratio from a 2x2 table}
\usage{
oddsratio_ci(a, b, c, d, level = 0.99)
}
\arguments{
\item{a, b, c, d}{Cell counts of the 2x2 contingency table}

\item{level}{Confidence level (default 0.99)}
}
\value{
Object of class 'oddsratio_ci'
}
\description{
Confidence interval for odds ratio from a 2x2 table
}
\examples{
oddsratio_ci(a = 12, b = 5, c = 4, d = 15)
oddsratio_ci(a = 12, b = 5, c = 4, d = 15, level = 0.95)
}
