% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_ci.R
\name{prop_ci}
\alias{prop_ci}
\title{Confidence interval for a proportion}
\usage{
prop_ci(trials, successes, level = 0.99)
}
\arguments{
\item{trials}{Number of trials}

\item{successes}{Number of successes}

\item{level}{Confidence level (default 0.99)}
}
\value{
Object of class 'prop_ci'
}
\description{
Confidence interval for a proportion
}
\examples{
# 45 successes out of 100 trials
prop_ci(trials = 100, successes = 45)
prop_ci(trials = 100, successes = 45, level = 0.95)
}
